requirements.txt
discord.py-self
aiohttp
rich
pyttsx3
